refactor: The table interface has been moved to a separate component
